Code for PAT1
The code is Overcomplicated. Do not follow this. Try to create a structure and use that instead.
There was a small error in my code which led to an infinity loop.
instead of ++*(top), I did *(++top).
It was nearly impossible to Debug this Code as Vpropel won't show you any output of your program whatsoever if you run into an infinity loop.
It will just say timeout, core dumped.